requirements.txt
alembic>=1.4.2
Flask>=1.1.2
Flask-Cors>=3.0.8
Flask-RESTful>=0.3.8
PyMySQL>=0.10.0
pyOpenSSL>=19.1.0
redis>=3.5.3
SQLAlchemy>=1.3.19
requests>=2.24.0
bcrypt>=3.2.0
python-redis-lock>=3.6.0
aliyun-python-sdk-core-v3>=2.13.11
aliyun-python-sdk-dysmsapi>=2.1.1
pika

pillow
